separar vistas
